auth payload validation and tests
